Made form more user friendly

+++ applications/ipos/controllers/edit.py
@@ -39,38 +39,41 @@ def export():
     return s.getvalue()
 
 def import_and_sync():
-    form = FORM(INPUT(_type='file', _name='data'), INPUT(_type='submit'))
-    if form.process().accepted:
-        db.import_from_csv_file(form.vars.data.file,unique=False)
+    if request.cookies.has_key('authenticate') and request.cookies['authenticate'].value == 'true':
+        form = FORM(INPUT(_type='file', _name='data'), INPUT(_type='submit'))
+        if form.process().accepted:
+            db.import_from_csv_file(form.vars.data.file,unique=False)
 
-        #delete all old sets of data sources. old means was not uploaded just now
-        exportTime = db(db.data_migration).select(db.data_migration.export_time, db.data_migration.uuid, orderby=~db.data_migration.modified_on).first().export_time
-        if exportTime: #If there's no export time, it's not coming from a migration, want to avoid doing anything
-            db(db.data_migration.export_time != exportTime).delete() # delete all migration rows that were not part of this export                 
+            #delete all old sets of data sources. old means was not uploaded just now
+            exportTime = db(db.data_migration).select(db.data_migration.export_time, db.data_migration.uuid, orderby=~db.data_migration.modified_on).first().export_time
+            if exportTime: #If there's no export time, it's not coming from a migration, want to avoid doing anything
+                db(db.data_migration.export_time != exportTime).delete() # delete all migration rows that were not part of this export                 
 
-            #Get the newest datasource
-            newDataSourceUUID = db(db.data_migration).select(db.data_migration.uuid,orderby=~db.data_migration.created_time).first().uuid
+                #Get the newest datasource
+                newDataSourceUUID = db(db.data_migration).select(db.data_migration.uuid,orderby=~db.data_migration.created_time).first().uuid
 
-            #Delete all data referencing old data_migration
-            for table in db.tables:
-                if 'uuid' in db[table].fields and ('data_migration_id' in db[table].fields or table =='data_migration'):
-                    if table != 'data_migration':
-                        #Delete all records that arent' referencing the most recent data_migration_id
-                        db(db[table].data_migration_id != newDataSourceUUID).delete()
+                #Delete all data referencing old data_migration
+                for table in db.tables:
+                    if 'uuid' in db[table].fields and ('data_migration_id' in db[table].fields or table =='data_migration'):
+                        if table != 'data_migration':
+                            #Delete all records that arent' referencing the most recent data_migration_id
+                            db(db[table].data_migration_id != newDataSourceUUID).delete()
 
-                    #For every uuid, delete all but the latest
-                    #This is precautionary, in case we happen to be uploading the exact same state of the db.  Avoids duplicates
-                    
-                    #Note both forms of deletion are necessary because they handle the cases in which a record was deleted, 
-                    #a record was added, a duplicate record has been imported
-                    
-                    items = db(db[table]).select(db[table].id, db[table].uuid, orderby=db[table].uuid | ~db[table].modified_on)
-                    if items and len(items) > 0:
-                        prevUuid = None
-                        for item in items:
-                            if item.uuid != prevUuid:
-                                prevUuid = item.uuid
-                                db((db[table].uuid==item.uuid) & (db[table].id!=item.id)).delete()
+                        #For every uuid, delete all but the latest
+                        #This is precautionary, in case we happen to be uploading the exact same state of the db.  Avoids duplicates
+                        
+                        #Note both forms of deletion are necessary because they handle the cases in which a record was deleted, 
+                        #a record was added, a duplicate record has been imported
+                        
+                        items = db(db[table]).select(db[table].id, db[table].uuid, orderby=db[table].uuid | ~db[table].modified_on)
+                        if items and len(items) > 0:
+                            prevUuid = None
+                            for item in items:
+                                if item.uuid != prevUuid:
+                                    prevUuid = item.uuid
+                                    db((db[table].uuid==item.uuid) & (db[table].id!=item.id)).delete()
+    else:
+        redirect(URL('default','matcher'))                           
 
     return dict(form=form)
 
@@ -87,12 +90,15 @@ def add_company():
         ipo_info.data_migration_id.writable = ipo_info.data_migration_id.readable = False
         ipo_info.company_id.writable = ipo_info.company_id.readable = False
         ipo_info.uuid.writable = ipo_info.uuid.readable = False
+        ipo_info.date_week.widget=SQLFORM.widgets.date.widget
 
         description = db.company_description
         description.company_id.writable =  description.company_id.readable = False
         description.data_migration_id.writable = description.data_migration_id.readable = False
         description.uuid.writable = description.uuid.readable = False
-        
+        description.description.widget=lambda field, value: SQLFORM.widgets.text.widget(field, value, _class="company_description_textarea")
+
+
         record = db((db.company_info.uuid == request.args(0)) & (db.ipo_info.company_id == request.args(0)) & (db.company_description.company_id == request.args(0))).select().first()
         if record:
             message = 'Update company Information'
@@ -102,7 +108,7 @@ def add_company():
                 for attribute in record[key]:
                     if key == "company_info" or attribute != 'id':
                         dictRecord[attribute] = record[key][attribute]
-            form=SQLFORM.factory(company_info,ipo_info, description, record=dictRecord, showid=False)
+            form=SQLFORM.factory(company_info,ipo_info, description, record=dictRecord, showid=False, _class='add_or_edit_company', formstyle='table2cols')
             if form.process().accepted:
                 form.vars.company_id=form.vars.uuid
                 id = record.company_info.update_record(**db.company_info._filter_fields(form.vars))
@@ -116,7 +122,12 @@ def add_company():
                 response.flash = 'please fill the form'
         else: 
             message = 'Add new company'
-            form=SQLFORM.factory(company_info,ipo_info, description)
+            form=SQLFORM.factory(company_info,ipo_info, description, _class='add_or_edit_company', formstyle='table2cols')
+            logger.info(form)
+            print form.element('input[name=description]')
+            # ['_type']='textarea'
+            logger.info("--------")
+            logger.info(form)
             if form.process().accepted:
                 id = db.company_info.insert(**db.company_info._filter_fields(form.vars))
                 form.vars.company_id = db(db.company_info.id == id).select(db.company_info.uuid).first().uuid #get the company's uuid
